moved execute function to cipper.js [ch7476]
